Improved filter
First, each row in the database table is looped through to get the number of ingredients in each recipe. Then the user's selected ingredients array is looped through to get the number of ingredients that match the ingredients in each recipe in the table. If the user selects all the ingredients that are in a recipe, that recipe will be output. It does not matter if the user selects even more ingredients than there are in a recipe as long as each ingredient in the recipe is selected.